@@ -1,6 +1,6 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFramework>netcoreapp3.1</TargetFramework>
+		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
 		<DefineConstants>NETCORE;</DefineConstants>
 		<DefineConstants>NETCOREAPP1_1</DefineConstants>
 		<PackageId>GeneXus.SDAPI.Core</PackageId>
