@@ -1,10 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFramework>netcoreapp3.1</TargetFramework>
+		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
 		<DefineConstants>NETCORE</DefineConstants>
 		<PackageTags>DynService OData</PackageTags>
 		<PackageId>GeneXus.DynService.OData.Core</PackageId>
 		<AssemblyName>GeneXus.Data.DynService.OData</AssemblyName>
+		<NoWarn>SYSLIB0014</NoWarn>
 	</PropertyGroup>
 
   <ItemGroup>
