@@ -1,12 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-	<DefineConstants>NETCORE;</DefineConstants>
-	<PackageTags>GXGeolocation GXHttpServices WebSocket</PackageTags>
-	<PackageId>GeneXus.Classes.Web.Core</PackageId>
-	<PackageTags>GXGeolocation GXHttpServices WebSocket</PackageTags>
-	<PackageId>GeneXus.Classes.Web.Core</PackageId>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+		<DefineConstants>NETCORE;</DefineConstants>
+		<PackageTags>GXGeolocation GXHttpServices WebSocket</PackageTags>
+		<PackageId>GeneXus.Classes.Web.Core</PackageId>
+		<PackageTags>GXGeolocation GXHttpServices WebSocket</PackageTags>
+		<PackageId>GeneXus.Classes.Web.Core</PackageId>
+		<NoWarn>CA1812,CA1819</NoWarn>
+	</PropertyGroup>
 
   <ItemGroup>
     <Compile Include="..\..\dotnetframework\GxClasses\Domain\GXGeolocation.cs" Link="Domain\GXGeolocation.cs" />
@@ -14,7 +15,6 @@
     <Compile Include="..\..\dotnetframework\GxClasses\Middleware\GXHttpServices.cs" Link="Middleware\GXHttpServices.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Domain\GXRuntime.cs" Link="Domain\GXRuntime.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Services\GxRestWrapper.cs" Link="Middleware\GxRestWrapper.cs" />
-    <Compile Include="..\..\dotnetframework\GxClasses\Services\JsonConverter.cs" Link="Helpers\JsonConverter.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Services\ReflectionHelper.cs" Link="Helpers\ReflectionHelper.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\View\GXGridStateHandler.cs" Link="View\GXGridStateHandler.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\View\GXWebControls.cs" Link="View\GXWebControls.cs" />
@@ -29,7 +29,6 @@
 
   <ItemGroup>
     <PackageReference Include="log4net" Version="2.0.11" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.0.0" />
   </ItemGroup>
 
   <ItemGroup>
