@@ -1,11 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-	<TargetFramework>netcoreapp3.1</TargetFramework>
-	<OutputType>Library</OutputType>
-    <DefineConstants>NETCORE</DefineConstants>
-	<PackageTags>Data Access</PackageTags>
-	<PackageId>GeneXus.Classes.Core</PackageId>
+	<PropertyGroup>
+		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+		<OutputType>Library</OutputType>
+		<DefineConstants>NETCORE</DefineConstants>
+		<PackageTags>Data Access</PackageTags>
+		<PackageId>GeneXus.Classes.Core</PackageId>
+		<EnableUnsafeBinaryFormatterSerialization>true</EnableUnsafeBinaryFormatterSerialization>
+		<NoWarn>618;1607;1698;SYSLIB0021;SYSLIB0027;SYSLIB0028;SYSLIB0023</NoWarn>
 	</PropertyGroup>
 
    <ItemGroup>
