@@ -1,6 +1,6 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFramework>netcoreapp3.1</TargetFramework>
+		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
 		<OutputType>Exe</OutputType>
 		<PackageTags>DataInitialization</PackageTags>
 		<PackageId>GeneXus.DataInitialization.Core</PackageId>
@@ -28,9 +28,13 @@
   </ItemGroup>
 
    <ItemGroup>
-		<Content Include="$(ProjectName).deps.json">
+		<Content Include="net6.0\$(MSBuildProjectName).deps.json">
 			<Pack>true</Pack>
-			<PackagePath>\content\</PackagePath>
+			<PackagePath>\content\net6.0\</PackagePath>
+		</Content>
+		<Content Include="net5.0\$(MSBuildProjectName).deps.json">
+			<Pack>true</Pack>
+			<PackagePath>\content\net5.0\</PackagePath>
 		</Content>
 	</ItemGroup>
 </Project>
