@@ -267,11 +267,11 @@ namespace GeneXus.Application
 				{
 					msglistItem msgItem = (msglistItem)msg[0];
 					if (msgItem.gxTpr_Id.Contains("NotFound"))
-						return SetError("404", msgItem.gxTpr_Description);
+						return SetError(HttpStatusCode.NotFound.ToString(HttpHelper.INT_FORMAT), msgItem.gxTpr_Description);
 					else if (msgItem.gxTpr_Id.Contains("WasChanged"))
-						return SetError("409", msgItem.gxTpr_Description);
+						return SetError(HttpStatusCode.Conflict.ToString(HttpHelper.INT_FORMAT), msgItem.gxTpr_Description);
 					else
-						return SetError("400", msgItem.gxTpr_Description);
+						return SetError(HttpStatusCode.BadRequest.ToString(HttpHelper.INT_FORMAT), msgItem.gxTpr_Description);
 				}
 			}
 			return Task.CompletedTask;
