@@ -32,9 +32,6 @@ namespace GeneXus.Utils
 	class NTSGeographyWrapper
 	{
 
-		private const double EARTH_RADIUS = 6371008.8;
-		private const double DEG_TO_RAD = Math.PI / 180;
-
 		internal static object Parse(String geoText)
 		{
 			WKTReader reader = new WKTReader();
@@ -63,6 +60,10 @@ namespace GeneXus.Utils
 			{
 				return null;
 			}
+		}
+		internal static bool IsNull(object instance)
+		{
+			return (instance == null || instance == NTSGeographyWrapper.NullSQLGeography);
 		}
 
 		internal static object STGeometryType(object instance)
@@ -150,27 +151,6 @@ namespace GeneXus.Utils
 			var dt = g.Inverse(LatA, LonA, LatB, LonB);
 			return dt.s12;
 		}
-
-		/*public static double DistanceSLC(double Lat1, double Lon1, double Lat2, double Lon2)
-		{
-			try
-			{
-				double radLat1 = Lat1 * DEG_TO_RAD;
-				double radLat2 = Lat2 * DEG_TO_RAD;
-				double radLon1 = Lon1 * DEG_TO_RAD;
-				double radLon2 = Lon2 * DEG_TO_RAD;
-
-				// central angle, aka arc segment angular distance
-				double centralAngle = Math.Acos(Math.Sin(radLat1) * Math.Sin(radLat2) +
-						Math.Cos(radLat1) * Math.Cos(radLat2) * Math.Cos(radLon2 - radLon1));
-
-				// great-circle (orthodromic) distance on Earth between 2 points
-				return EARTH_RADIUS * centralAngle;
-			}
-			catch {
-					throw;
-				}
-		}*/
 
 		internal static bool STIntersects(object instanceA, object instanceB)
 		{
@@ -673,22 +653,27 @@ namespace GeneXus.Utils
 			}
 			try
 			{
-				// Parse
 				_innerValue = NTSGeographyWrapper.Parse(geoText);
 				// SRID, Type & Points X, Y
-				if ((!NTSGeographyWrapper.IsValid(_innerValue)) && _innerValue != null)
+				if (_innerValue != null && !NTSGeographyWrapper.IsNull(_innerValue))
 				{
-					//_innerValue = NTSGeographyWrapper.MakeValid(_innerValue);
-				}
+					if (NTSGeographyWrapper.IsValid(_innerValue))
+					{
+						//_innerValue = NTSGeographyWrapper.MakeValid(_innerValue);	
+						this.srid = NTSGeographyWrapper.Srid(_innerValue);
+						this.setGXGeoType(NTSGeographyWrapper.STGeometryType(_innerValue).ToString());
 
-				this.srid = NTSGeographyWrapper.Srid(_innerValue);
-
-				this.setGXGeoType(NTSGeographyWrapper.STGeometryType(_innerValue).ToString());
-				if (GeographicType == GeoGraphicTypeValue.Point)
-				{
-					this.Point.Longitude = NTSGeographyWrapper.Long(_innerValue);
-					this.Point.Latitude = NTSGeographyWrapper.Lat(_innerValue);
+						if (GeographicType == GeoGraphicTypeValue.Point)
+						{
+							this.Point.Longitude = NTSGeographyWrapper.Long(_innerValue);
+							this.Point.Latitude = NTSGeographyWrapper.Lat(_innerValue);
+						}
+					}
+					else
+						setNullGeography();
 				}
+				else
+					setNullGeography();
 			}
 			catch (Exception ex)
 			{
@@ -700,22 +685,24 @@ namespace GeneXus.Utils
 					}
 					else
 					{
-						// Cannot parse value
-						_innerValue = NTSGeographyWrapper.NullSQLGeography;
-						this.geoText = "";
-						this.Point.Longitude = 0;
-						this.Point.Latitude = 0;
+						setNullGeography();
 					}
 				}
 				else
 				{
-					// Cannot parse value
-					_innerValue = NTSGeographyWrapper.NullSQLGeography;
-					this.geoText = "";
-					this.Point.Longitude = 0;
-					this.Point.Latitude = 0;
+					setNullGeography();
+					
 				}
 			}
+		}
+
+		void setNullGeography()
+		{
+			// Cannot parse value
+			_innerValue = NTSGeographyWrapper.NullSQLGeography;
+			this.geoText = "";
+			this.Point.Longitude = 0;
+			this.Point.Latitude = 0;
 		}
 
 		override public String ToString()
@@ -954,7 +941,6 @@ namespace GeneXus.Utils
 			}
 			return ArrayOfPoints;
 		}
-
 	}
 
 	public class PointT
@@ -972,6 +958,5 @@ namespace GeneXus.Utils
 	{
 		public Line[] Points;
 	}
-
 
 }
