@@ -69,6 +69,7 @@ namespace GeneXus.Data.NTier
         int RecordCount { get;}
         void OnCommitEvent(object instance, string method);
         int readNextErrorRecord();
+		List<ParDef> DynamicParameters { get; }
     }
 	public interface IFieldGetter
 	{
@@ -146,7 +147,7 @@ namespace GeneXus.Data.NTier
 
 		public virtual string getDataStoreName()
 		{
-			return "Default";
+			return Preferences.DefaultDatastore;
 		}
 		public void setParameters(int cursor,
 							   IFieldSetter stmt,
@@ -168,7 +169,7 @@ namespace GeneXus.Data.NTier
 					else if (pdef.Out)
 					{
 						stmt.RegisterOutParameter(idxParmCollection, null);
-						continue;
+						goto Increment;
 					}
 
 					if (pdef.Nullable)
@@ -254,6 +255,7 @@ namespace GeneXus.Data.NTier
 								break;
 						}
 					}
+				Increment:
 					idx += 1;
 					idxParmCollection += 1;
 				}
@@ -417,7 +419,8 @@ namespace GeneXus.Data.NTier
 
 						if (parmHasValue != null)
 						{
-							if (oCur.getFieldSetter().ParameterDefinition.Count == 0) //Backward compatibility
+							List<ParDef> pdefList = oCur.DynamicParameters;
+							if (pdefList.Count == 0) //Backward compatibility
 							{
 								Object[] parmsNew = new Object[parms.Length + parmHasValue.Length];
 								parmHasValue.CopyTo(parmsNew, 0);
@@ -427,12 +430,23 @@ namespace GeneXus.Data.NTier
 							else
 							{
 								List<object> parmsNew = new List<object>();
-								for (int i=0; i<parms.Length; i++)
+								int idx = 0;
+								for (int i=0; i< pdefList.Count; i++)
 								{
-									if ((short)parmHasValue[i] == 0)
+									ParDef pdef = pdefList[i];
+									if (pdef.Nullable)
 									{
-										parmsNew.Add(parms[i]);
+										if (parmHasValue[i] == 0)
+										{
+											parmsNew.Add(parms[idx]);
+										}
+										idx += 1;
 									}
+									if (parmHasValue[i] == 0)
+									{
+										parmsNew.Add(parms[idx]);
+									}
+									idx += 1;
 								}
 								_dataStoreHelper.setParameters(cursor, oCur.getFieldSetter(), parmsNew.ToArray());
 							}
