@@ -550,6 +550,7 @@ namespace GeneXus.Data
 		static readonly ILog log = log4net.LogManager.GetLogger(typeof(GxODPManagedOracle));
 		static Assembly _odpAssembly;
 		const string OracleDbTypeEnum = "Oracle.ManagedDataAccess.Client.OracleDbType";
+		const string OracleAssemblyName = "Oracle.ManagedDataAccess";
 
 		public static Assembly OdpAssembly
 		{
@@ -559,11 +560,11 @@ namespace GeneXus.Data
 				{
 					if (_odpAssembly == null)
 					{
-						string assemblyPath = Path.Combine(FileUtil.GetStartupDirectory(), "Oracle.ManagedDataAccess.dll");
+						string assemblyPath = Path.Combine(FileUtil.GetStartupDirectory(), $"{OracleAssemblyName}.dll");
 						GXLogging.Debug(log, "Loading Oracle.ManagedDataAccess from:" + assemblyPath);
 #if NETCORE
 						var asl = new AssemblyLoader(FileUtil.GetStartupDirectory());
-						_odpAssembly = asl.LoadFromAssemblyPath(assemblyPath);
+						_odpAssembly = asl.LoadFromAssemblyName(new AssemblyName(OracleAssemblyName));
 #else
 						if (File.Exists(assemblyPath))
 						{
@@ -575,7 +576,7 @@ namespace GeneXus.Data
 							_odpAssembly = Assembly.LoadWithPartialName("Oracle.ManagedDataAccess");
 						}
 #endif
-						GXLogging.Debug(log, "Oracle.ManagedDataAccess Loaded:" + _odpAssembly.FullName + " location: " + _odpAssembly.Location + " CodeBase:" + _odpAssembly.CodeBase);
+						GXLogging.Debug(log, "Oracle.ManagedDataAccess Loaded:" + _odpAssembly.FullName + " location: " + _odpAssembly.Location);
 					} 
 
 				}
@@ -1278,7 +1279,7 @@ namespace GeneXus.Data
 		}
 		public override Object Net2DbmsGeo(GXType type, IGeographicNative geo)
 		{
-			return geo.ToStringSQL();
+			return geo.ToStringSQL("POINT(0 0)");
 		}
 
 		private static readonly string[] ConcatOpValues = new string[] { string.Empty, " || ", string.Empty };
