@@ -209,7 +209,7 @@ public class GxExternalDirectoryInfo : IGxDirectoryInfo
             searchPattern = "";
         List<string> files = _provider.GetFiles(_name, searchPattern);
 		
-		GxExternalFileInfo[] externalFiles = files.Select(elem => new GxExternalFileInfo(elem, _provider, GxFileType.Public)).ToArray();
+		GxExternalFileInfo[] externalFiles = files.Select(elem => new GxExternalFileInfo(elem, _provider, GxFileType.Default)).ToArray();
         return externalFiles;
     }
 
@@ -412,41 +412,41 @@ public class GxExternalFileInfo : IGxFileInfo
     private string _name;
 	private ExternalProvider _provider;
     private string _url;	
-	private GxFileType _fileTypeAtt = GxFileType.Public;
+	private GxFileType _fileTypeAtt = GxFileType.Private;
 
 	public GxExternalFileInfo(ExternalProvider provider)
     {
         _provider = provider;
-        _name = "";
-        _url = "";
+        _name = string.Empty;
+        _url = string.Empty;
     }
 
-    public GxExternalFileInfo(string storageObjectFullname, ExternalProvider provider, GxFileType fileType)
+    public GxExternalFileInfo(string objectPath, ExternalProvider provider, GxFileType fileType)
     {
-		storageObjectFullname = storageObjectFullname!=null ? storageObjectFullname.Replace('\\', '/') : storageObjectFullname;
-		_name = storageObjectFullname;
+		objectPath = !String.IsNullOrEmpty(objectPath) ? objectPath.Replace('\\', '/') : objectPath;
         _provider = provider;
+		_fileTypeAtt = fileType;
+		_name = objectPath;
+
 		Uri result;
-		if (Uri.TryCreate(storageObjectFullname, UriKind.Absolute, out result) && result.IsAbsoluteUri)
+		if (Uri.TryCreate(objectPath, UriKind.Absolute, out result) && result.IsAbsoluteUri)
 		{
-			_url = storageObjectFullname;
+			_url = objectPath;
 		}
 		else {
-			if (fileType.HasFlag(GxFileType.Attribute)) //Attributes multimedia consider Storage Provider Folder 
+			string folderName = ((ExternalProviderBase)provider).Folder;
+			if (!string.IsNullOrEmpty(folderName) && fileType.HasFlag(GxFileType.Attribute) && !_name.StartsWith(folderName))
 			{
-				_url = provider.GetBaseURL() + storageObjectFullname;
-				_name = _url.Replace(provider.StorageUri, string.Empty);
-				if (_name.StartsWith("/"))
-					_name = _name.Substring(1, _name.Length - 1);
+				_url = $"{provider.GetBaseURL()}{_name}";
+				_name = $"{folderName}{StorageUtils.DELIMITER}{_name}";
 			}
-		}		
-		_fileTypeAtt = fileType;
+		}
 	}
 
-    public GxExternalFileInfo(string storageObjectFullname, string url, ExternalProvider provider, GxFileType fileType = GxFileType.Public)
-    {
-        _name = storageObjectFullname;
-        _provider = provider;
+    public GxExternalFileInfo(string objectPath, string url, ExternalProvider provider, GxFileType fileType = GxFileType.Private)
+    {				
+        _name = StorageFactory.GetProviderObjectAbsoluteUriSafe(provider, objectPath);
+		_provider = provider;
         _url = url;
 		_fileTypeAtt = fileType;
 	}
@@ -576,12 +576,8 @@ public class GxExternalFileInfo : IGxFileInfo
 
 	private string URL
 	{
-		get {
-			if (string.IsNullOrEmpty(_url))
-			{
-				_url = _provider.Get(_name, _fileTypeAtt, 0);
-			}
-			return _url;
+		get {			
+			return _provider.GetUrl(_name, _fileTypeAtt, 0);
 		}
 	}
 
@@ -640,11 +636,12 @@ public class GxExternalFileInfo : IGxFileInfo
 [Flags]
 public enum GxFileType
 {
-	Public = 0,
-	Private = 1,
-	Attribute = 2, 
-	PublicAttribute = Attribute | Public,
-	PrivateAttribute = Attribute | Private,
+	Default = 0,
+	PublicRead = 1,
+	Private = 2,
+	Attribute = 8,
+	DefaultAttribute = Attribute | Default,
+	PrivateAttribute = Attribute | Private
 }
 
 public class GxFile
@@ -670,13 +667,13 @@ public class GxFile
             _baseDirectory = _baseDirectory.Substring(0, _baseDirectory.Length - 1);
     }
 
-    public GxFile(string baseDirectory, IGxFileInfo file, GxFileType fileType = GxFileType.Public)
+    public GxFile(string baseDirectory, IGxFileInfo file, GxFileType fileType = GxFileType.Private)
       : this(baseDirectory)
     {
         _file = file;
     }
 
-    public GxFile(string baseDirectory, string fileName, GxFileType fileType = GxFileType.Public)
+    public GxFile(string baseDirectory, string fileName, GxFileType fileType = GxFileType.Private)
       : this(baseDirectory)
     {
 		if (GxUploadHelper.IsUpload(fileName))
@@ -733,22 +730,26 @@ public class GxFile
 					}
 					else
 					{
+						_file = null;
 						if (GxUploadHelper.IsUpload(value))
 						{
 							_uploadFileId = value;
 							value = GxUploadHelper.UploadPath(value);
+							ExternalProvider provider = ServiceFactory.GetExternalProvider();
+							_file = (provider != null)? new GxExternalFileInfo(value, String.Empty, provider): _file;
 						}
-						
-						if (IsAbsoluteUrl(value))
+
+						if (_file == null)
 						{
-							string objName = string.Empty;
-							ExternalProvider p = StorageFactory.GetExternalProviderFromUrl(value, out objName);
-							_file = new GxExternalFileInfo(objName, value, p);
-						}
-						else
-						{
-							_file = new GxFileInfo(_baseDirectory);
-							_file.Source = value;
+							if (IsAbsoluteUrl(value))
+							{
+								_file =new GxExternalFileInfo(value, value, ServiceFactory.GetExternalProvider());
+							}
+							else
+							{
+								_file = new GxFileInfo(_baseDirectory);
+								_file.Source = value;
+							}
 						}
 
 						_lastError = 0;
