@@ -372,40 +372,47 @@ namespace GeneXus.Utils
 		}
 		public void Add(Object o, int idx)
 		{
-			T TObject;
-			if (typeof(T).IsAssignableFrom(o.GetType()))
-				TObject = (T)o;
-			else if (typeof(IGxJSONAble).IsAssignableFrom(typeof(T)))
-			{
-
-				TObject = (T)Activator.CreateInstance(typeof(T));
-				((IGxJSONAble)TObject).FromJSONObject(o);
-			}
-			else
-			{
-				if (typeof(T) == typeof(Geospatial))
-				{
-					object g = (Geospatial)(string)o;
-					TObject = (T)g;
-				}
-				else if (typeof(T) == typeof(Guid))
-				{
-					object g = new Guid(o.ToString());
-					TObject = (T)g;
-				}
-				else if (o is IConvertible)
-				{
-					TObject = (T)Convert.ChangeType(o, typeof(T));
-				}
-				else
-					TObject = (T)Convert.ChangeType(o.ToString(), typeof(T));
-			}
+			T TObject = ConvertToT(o);
 			if (idx == 0)
 				Add(TObject);
 			else
 				Insert(idx - 1, TObject);
 		}
+		private T ConvertToT(Object o)
+		{
+			T TObject=default(T);
+			if (o != null)
+			{
+				if (typeof(T).IsAssignableFrom(o.GetType()))
+					TObject = (T)o;
+				else if (typeof(IGxJSONAble).IsAssignableFrom(typeof(T)))
+				{
 
+					TObject = (T)Activator.CreateInstance(typeof(T));
+					((IGxJSONAble)TObject).FromJSONObject(o);
+				}
+				else
+				{
+					if (typeof(T) == typeof(Geospatial))
+					{
+						object g = (Geospatial)(string)o;
+						TObject = (T)g;
+					}
+					else if (typeof(T) == typeof(Guid))
+					{
+						object g = new Guid(o.ToString());
+						TObject = (T)g;
+					}
+					else if (o is IConvertible)
+					{
+						TObject = (T)Convert.ChangeType(o, typeof(T));
+					}
+					else
+						TObject = (T)Convert.ChangeType(o.ToString(), typeof(T));
+				}
+			}
+			return TObject;
+		}
 		public void RemoveItem(int idx)
 		{
 			if (idx <= 0 || idx > Count)
@@ -430,7 +437,8 @@ namespace GeneXus.Utils
 		}
 		public int IndexOf(object value)
 		{
-			return base.IndexOf((T)value) + 1;
+			T TObject = ConvertToT(value);
+			return base.IndexOf(TObject) + 1;
 		}
 
 		public virtual void writexml(GXXMLWriter oWriter, string sName)
@@ -1005,6 +1013,21 @@ namespace GeneXus.Utils
 		public short opened;
 	}
 
+	[AttributeUsage(AttributeTargets.Class)]
+	public sealed class GxJsonName : Attribute
+	{
+		string _jsonName;
+		public GxJsonName(String name)
+		{
+			_jsonName = name;
+		}
+		public string Name
+		{
+			get { return this._jsonName;  }
+			set { this._jsonName = value; }
+		}
+	}
+
 	[Serializable]
 	[XmlType(IncludeInSchema = false)]
 	public class GxUserType : IGxXMLSerializable, ICloneable, IGxJSONAble, IGxJSONSerializable
@@ -1305,7 +1328,7 @@ namespace GeneXus.Utils
 					tagCloseIndex--;
 				}
 
-				if (!string.IsNullOrEmpty(defaultNamespace) || forceDefaultNamespace || prefixes.Keys.Contains(""))
+				if (!string.IsNullOrEmpty(defaultNamespace) || forceDefaultNamespace || prefixes.ContainsKey(""))
 				{
 					string currentTagSubstring = xml.Substring(tagOpenIndex, tagCloseIndex - tagOpenIndex);
 					int nsIndex = currentTagSubstring.IndexOf("xmlns=");
@@ -1577,7 +1600,7 @@ namespace GeneXus.Utils
 									MethodInfo setBlob = GetMethodInfo("gxtv_" + GetType().Name + "_" + name + "_setblob");
 									if (setBlob != null)
 									{
-										if (HttpHelper.GetHttpRequestPostedFile(context.HttpContext, sVar, out uploadPath))
+										if (HttpHelper.GetHttpRequestPostedFile(context, sVar, out uploadPath))
 										{
 											string fileName = HttpHelper.GetHttpRequestPostedFileName(this.context.HttpContext, sVar);
 											string fileType = HttpHelper.GetHttpRequestPostedFileType(this.context.HttpContext, sVar);
