@@ -5,8 +5,10 @@ using System.IO;
 using System.Runtime.Serialization.Json;
 using System.Text;
 using Jayrock.Json;
-#if NETCORE
+using System.Runtime.Serialization;
+using GeneXus.Configuration;
 using System.Linq;
+#if NETCORE
 using System.Buffers.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -27,7 +29,7 @@ namespace GeneXus.Utils
 			switch (reader.TokenType)
 			{
 				case JsonTokenType.True:
-					return false;
+					return true;
 				case JsonTokenType.False:
 					return false;
 				case JsonTokenType.StartArray:
@@ -136,6 +138,7 @@ namespace GeneXus.Utils
 	{
 		
 		static readonly ILog log = log4net.LogManager.GetLogger(typeof(GeneXus.Utils.JSONHelper));
+		static string WFCDateTimeFormat = Preferences.WFCDateTimeMillis ? DateTimeUtil.JsonDateFormatMillis : DateTimeUtil.JsonDateFormat;
 		public static bool IsJsonNull(object jobject)
 		{
 			return GXJsonSerializer.Instance.IsJsonNull(jobject);
@@ -214,8 +217,7 @@ namespace GeneXus.Utils
 		{
 			try
 			{
-				var settings = new DataContractJsonSerializerSettings() { DateTimeFormat = new System.Runtime.Serialization.DateTimeFormat(DateTimeUtil.JsonDateFormatMillis), KnownTypes = knownTypes };
-				
+				var settings = SerializationSettings(knownTypes);
 				DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(T), settings);
 				
 				using (MemoryStream stream = new MemoryStream())
@@ -229,6 +231,44 @@ namespace GeneXus.Utils
 				GXLogging.Error(log, "Serialize error ", ex);
 			}
 			return null;
+		}
+		internal static string WCFSerialize<T>(T kbObject, Encoding encoding, IEnumerable<Type> knownTypes, bool useSimpleDictionaryFormat) where T : class
+		{
+			try
+			{
+				var settings = WCFSerializationSettings(knownTypes, useSimpleDictionaryFormat);
+				DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(T), settings);
+				using (MemoryStream stream = new MemoryStream())
+				{
+					serializer.WriteObject(stream, kbObject);
+					return encoding.GetString(stream.ToArray());
+				}
+			}
+			catch (Exception ex)
+			{
+				GXLogging.Error(log, "Serialize error ", ex);
+			}
+			return null;
+		}
+		internal static void WCFSerialize<T>(T kbObject, Encoding encoding, IEnumerable<Type> knownTypes, Stream stream) where T : class
+		{
+			try
+			{
+				var settings = WCFSerializationSettings(knownTypes);
+				DataContractJsonSerializer serializer = new DataContractJsonSerializer(typeof(T), settings);
+				serializer.WriteObject(stream, kbObject);
+			}
+			catch (Exception ex)
+			{
+				GXLogging.Error(log, "Serialize error ", ex);
+			}
+		}
+		static DataContractJsonSerializerSettings SerializationSettings(IEnumerable<Type> knownTypes)
+		{
+			return new DataContractJsonSerializerSettings() { DateTimeFormat = new DateTimeFormat(DateTimeUtil.JsonDateFormatMillis), KnownTypes=knownTypes };
+		}
+		static DataContractJsonSerializerSettings WCFSerializationSettings(IEnumerable<Type> knownTypes, bool useSimpleDictionaryFormat=false) {
+			return new DataContractJsonSerializerSettings() { DateTimeFormat = new DateTimeFormat(WFCDateTimeFormat), EmitTypeInformation = EmitTypeInformation.Never, UseSimpleDictionaryFormat= useSimpleDictionaryFormat, KnownTypes=knownTypes };
 		}
 		public static T Deserialize<T>(string kbObject, Encoding encoding, IEnumerable<Type> knownTypes) where T : class, new()
 		{
