@@ -227,7 +227,7 @@ namespace GeneXus.Http
 		protected void setEventMetadata(string EventName, string Metadata)
 		{
 			if (EventsMetadata[EventName] == null)
-				EventsMetadata[EventName] = "";
+				EventsMetadata[EventName] = string.Empty;
 			EventsMetadata[EventName] += Metadata;
 		}
 
@@ -366,7 +366,7 @@ namespace GeneXus.Http
 				if (objMessage.Contains("row"))
 					row = (string)objMessage["row"];
 				else
-					row = "";
+					row = string.Empty;
 				if (objMessage.Contains("gxstate"))
 				{
 					ParseGXStateParms((JObject)objMessage["gxstate"]);
@@ -713,10 +713,10 @@ namespace GeneXus.Http
 									JArray columnHashes = new JArray();
 									JArray hideCodeValues;
 									JArray AllCollData = value as JArray;
-									string Picture = parm.Contains("pic") ? (string)parm["pic"] : "";
+									string Picture = parm.Contains("pic") ? (string)parm["pic"] : string.Empty;
 									if (parm.Contains("grid"))
 									{
-										string parentRow = "";
+										string parentRow = string.Empty;
 										//Case for each line command or collection based grid
 										if (AllCollData != null)
 										{
@@ -774,15 +774,15 @@ namespace GeneXus.Http
 													}
 												}
 												int gridId = (int)parm["grid"];
-												string pRowRCSuffix = (!String.IsNullOrEmpty(parentRow)) ? $"_{parentRow}" : String.Empty;
+												string pRowRCSuffix = (!String.IsNullOrEmpty(parentRow)) ? $"_{parentRow}" : string.Empty;
 												targetObj.FormVars[$"{cmpContext}nRC_GXsfl_{gridId.ToString(CultureInfo.InvariantCulture)}{pRowRCSuffix}"] = columnValues.Count.ToString(CultureInfo.InvariantCulture);
 											}
 										}
 										if (parm.Contains("prop") && ((string)parm["prop"]) == "GridRC")
 										{
-											string sRC = "";
-											string rowsufix = "";
-											string varname = "";
+											string sRC = string.Empty;
+											string rowsufix = string.Empty;
+											string varname = string.Empty;
 											if (jValue != null)
 											{
 												sRC = (string)(jValue["gridRC"]);
@@ -801,8 +801,8 @@ namespace GeneXus.Http
 											try
 											{
 												JObject hashObj = (JObject)(hash_i < inHashValues.Length ? inHashValues[hash_i] : new Jayrock.Json.JObject());
-												string sRow = hashObj.Contains("row") ? (string)hashObj["row"] : "";
-												string hash = hashObj.Contains("hsh") ? (string)hashObj["hsh"] : "";
+												string sRow = hashObj.Contains("row") ? (string)hashObj["row"] : string.Empty;
+												string hash = hashObj.Contains("hsh") ? (string)hashObj["hsh"] : string.Empty;
 												SetScalarOrCollectionValue((string)parm["av"], inParmsValues[parm_i], columnValues);
 												object TypedValue = getFieldValue(targetObj, (string)parm["av"]);
 												CheckParmIntegrity(TypedValue, hash, sRow, inParmsMetadata[parm_i], hash_i, Picture);
@@ -1000,8 +1000,8 @@ namespace GeneXus.Http
 		protected virtual bool IntegratedSecurityEnabled { get { return false; } }
 		protected virtual GAMSecurityLevel IntegratedSecurityLevel { get { return 0; } }
 		[Obsolete("IntegratedSecurityPermissionName is deprecated, it is here for compatibility. Use ExecutePermissionPrefix instead.", false)]
-		protected virtual string IntegratedSecurityPermissionName { get { return ""; } }
-		protected virtual string ExecutePermissionPrefix { get { return ""; } }
+		protected virtual string IntegratedSecurityPermissionName { get { return string.Empty; } }
+		protected virtual string ExecutePermissionPrefix { get { return string.Empty; } }
 		public bool IntegratedSecurityEnabled2 { get { return IntegratedSecurityEnabled; } }
 		public GAMSecurityLevel IntegratedSecurityLevel2 { get { return IntegratedSecurityLevel; } }
 #endif
@@ -1162,7 +1162,7 @@ namespace GeneXus.Http
 
 		public void AddThemeStyleSheetFile(String kbPrefix, String styleSheet, string urlBuildNumber)
 		{
-			string cssContent = "";
+			string cssContent = string.Empty;
 			Boolean bHasCustomContent = FetchCustomCSS(ref cssContent);
 
 			if (bHasCustomContent && !context.StyleSheetAdded(GetPgmname()))
@@ -1316,7 +1316,7 @@ namespace GeneXus.Http
 			{
 				if (staticContentBase == null)
 				{
-					string dir = "";
+					string dir = string.Empty;
 					if (Config.GetValueOf("STATIC_CONTENT", out dir))
 					{
 						if (!(dir.EndsWith("/") || dir.EndsWith("\\")) && !String.IsNullOrEmpty(dir))
@@ -1326,7 +1326,7 @@ namespace GeneXus.Http
 					}
 					else
 					{
-						staticContentBase = "";
+						staticContentBase = string.Empty;
 					}
 				}
 				return staticContentBase;
@@ -1389,36 +1389,51 @@ namespace GeneXus.Http
 			return Encrypt64(GXUtil.GetHash(WebSecurityHelper.StripInvalidChars(value), Cryptography.Constants.SecurityHashAlgorithm), key);
 		}
 
-		protected String Encrypt64(String value, String key)
+		protected string Encrypt64(string value, string key)
 		{
-			String sRet = "";
+			return Encrypt64(value, key, false);
+		}
+		private string Encrypt64(string value, string key, bool safeEncoding)
+		{
+			string sRet = string.Empty;
 			try
 			{
-				sRet = Crypto.Encrypt64(value, key);
+				sRet = Crypto.Encrypt64(value, key, safeEncoding);
 			}
 			catch (InvalidKeyException)
 			{
-				context.SetCookie("GX_SESSION_ID", "", "", DateTime.MinValue, "", context.GetHttpSecure());
+				context.SetCookie("GX_SESSION_ID", string.Empty, string.Empty, DateTime.MinValue, string.Empty, context.GetHttpSecure());
 				GXLogging.Error(log, "440 Invalid encryption key");
 				SendResponseStatus(440, "Session timeout");
 			}
 			return sRet;
 		}
-
-		protected String Decrypt64(String value, String key)
+		protected string UriEncrypt64(string value, string key)
 		{
-			String sRet = "";
+			return Encrypt64(value, key, true);
+		}
+		protected string Decrypt64(string value, string key)
+		{
+			return Decrypt64(value, key, false);
+		}
+		private string Decrypt64(string value, string key, bool safeEncoding)
+		{
+			String sRet = string.Empty;
 			try
 			{
-				sRet = Crypto.Decrypt64(value, key);
+				sRet = Crypto.Decrypt64(value, key, safeEncoding);
 			}
 			catch (InvalidKeyException)
 			{
-				context.SetCookie("GX_SESSION_ID", "", "", DateTime.MinValue, "", context.GetHttpSecure());
+				context.SetCookie("GX_SESSION_ID", string.Empty, string.Empty, DateTime.MinValue, string.Empty, context.GetHttpSecure());
 				GXLogging.Error(log, "440 Invalid encryption key");
 				SendResponseStatus(440, "Session timeout");
 			}
 			return sRet;
+		}
+		protected string UriDecrypt64(string value, string key)
+		{
+			return Decrypt64(value, key, true);
 		}
 
 		protected string DecryptAjaxCall(string encrypted)
@@ -1602,7 +1617,7 @@ namespace GeneXus.Http
 			{
 				strValue = ((Decimal)Value).ToString(CultureInfo.InvariantCulture);
 				if (strValue.IndexOf('.') != -1)
-					strValue = rgx2.Replace(rgx.Replace(strValue, ""), "");
+					strValue = rgx2.Replace(rgx.Replace(strValue, string.Empty), string.Empty);
 			}
 			else
 			{
@@ -1619,7 +1634,7 @@ namespace GeneXus.Http
 					{
 						strValue = ((Double)Value).ToString(CultureInfo.InvariantCulture);
 						if (strValue.IndexOf('.') != -1)
-							strValue = rgx2.Replace(rgx.Replace(strValue, ""), "");
+							strValue = rgx2.Replace(rgx.Replace(strValue, string.Empty), string.Empty);
 					}
 					else
 					{
@@ -1763,7 +1778,9 @@ namespace GeneXus.Http
 		public virtual bool UseBigStack()
 		{ return false; }
 
-		private const int STACKSIZE = 1024 * 1024 * 2;
+#if !NETCORE
+	private const int STACKSIZE = 1024 * 1024 * 2;
+#endif
 
 		public bool IsMain
 		{
@@ -1908,7 +1925,7 @@ namespace GeneXus.Http
 				String token = localHttpContext.Request.Headers["Authorization"];
 				if (!string.IsNullOrEmpty(token))
 				{
-					token = token.Replace("OAuth ", "");
+					token = token.Replace("OAuth ", string.Empty);
 					GxSecurityProvider.Provider.checkaccesstoken(context, token, out isOK);
 				}
 				else
@@ -2228,7 +2245,7 @@ namespace GeneXus.Http
 
 		public virtual string getresponse(string sGXDynURL)
 		{
-			return "";
+			return string.Empty;
 		}
 		public virtual void setparameters(object[] parms)
 		{
@@ -2243,7 +2260,7 @@ namespace GeneXus.Http
 			if (_currParameter < _params.Count)
 				return _params[_currParameter];
 			else
-				return "";
+				return string.Empty;
 		}
 		public string GetPar(string parameterName)
 		{
@@ -2288,7 +2305,7 @@ namespace GeneXus.Http
 			string sValue = null;
 			try
 			{
-				if (GxContext.GetHttpRequestPostedFile(localHttpContext, sVar, out sValue))
+				if (GxContext.GetHttpRequestPostedFile(context, sVar, out sValue))
 					return sValue;
 
 				if (FormVars != null)
@@ -2556,7 +2573,7 @@ namespace GeneXus.Http
 				{
 					UriVar = cgiGet(InternalName + "_GXI");
 				}
-				BlobVar = "";
+				BlobVar = string.Empty;
 			}
 		}
 
@@ -2582,7 +2599,7 @@ namespace GeneXus.Http
 						return target;
 				}
 			}
-			return "";
+			return string.Empty;
 		}
 
 		public void CallWebObject(string url)
@@ -2787,7 +2804,7 @@ namespace GeneXus.Http
 		{
 			context.WriteHtmlText(string.Format("ERROR: {0} is not a web component or could not load {0}.dll ", _DllName));
 		}
-		public override String getstring(String s) { return ""; }
+		public override String getstring(String s) { return string.Empty; }
 		public override void componentdrawstyles() { }
 		public override void componentprocess(string sPPrefix, string sPSFPrefix, string sEvt) { }
 		public override void componentprocess(string sPPrefix, string sPSFPrefix) { }
@@ -2805,7 +2822,7 @@ namespace GeneXus.Http
 		public GXNullWebComponent() { }
 		public override void componentstart() { }
 		public override void componentdraw() { }
-		public override String getstring(String s) { return ""; }
+		public override String getstring(String s) { return string.Empty; }
 		public override void componentdrawstyles() { }
 		public override void componentprocess(string sPPrefix, string sPSFPrefix, string sEvt) { }
 		public override void componentprocess(string sPPrefix, string sPSFPrefix) { }
@@ -2816,7 +2833,7 @@ namespace GeneXus.Http
 		public override void initialize() { }
 		public override string Name
 		{
-			get { return ""; }
+			get { return string.Empty; }
 		}
 		protected override void createObjects() { }
 	}
@@ -2948,10 +2965,10 @@ namespace GeneXus.Http
 		GXRadio meta = new GXRadio();
 		GXRadio metaequiv = new GXRadio();
 		GxStringCollection jscriptsrc = new GxStringCollection();
-		string caption = "";
+		string caption = string.Empty;
 		int backcolor;
 		int textcolor;
-		string background = "";
+		string background = string.Empty;
 		int visible;
 		int windowstate;
 		int enabled;
@@ -2959,11 +2976,11 @@ namespace GeneXus.Http
 		int left;
 		int width;
 		int height;
-		string internalname = "";
-		string bitmap = "";
-		string tag = "";
-		string _class = "";
-		string headerrawhtml = "";
+		string internalname = string.Empty;
+		string bitmap = string.Empty;
+		string tag = string.Empty;
+		string _class = string.Empty;
+		string headerrawhtml = string.Empty;
 
 		public GXRadio Meta
 		{
@@ -3065,7 +3082,7 @@ namespace GeneXus.Http
 
 		public static void AddResponsiveMetaHeaders(GXRadio meta)
 		{
-			TryAddMetaHeader(meta, "viewport", "width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no");
+			TryAddMetaHeader(meta, "viewport", "width=device-width, initial-scale=1, maximum-scale=4");
 			TryAddMetaHeader(meta, "apple-mobile-web-app-capable", "yes");
 		}
 
