@@ -1,24 +1,38 @@
 using GeneXus.Services;
+using GeneXus.Utils;
 
 namespace GeneXus.Storage
 {
 	public class StorageFactory
 	{
 		const char QUESTION_MARK = '?';
-		public static ExternalProvider GetExternalProviderFromUrl(string url, out string objectName)
+		public static bool TryGetProviderObjectName(ExternalProvider provider, string objectNameOrUrl, out string providerObjectName)
 		{
-			objectName = null;
-			ExternalProvider provider = ServiceFactory.GetExternalProvider();
-			if (provider != null)
+			providerObjectName = null;
+			if (provider != null && provider.TryGetObjectNameFromURL(objectNameOrUrl, out providerObjectName))
 			{
-				if (provider.GetObjectNameFromURL(url, out objectName))
+				int idx = providerObjectName.IndexOf(QUESTION_MARK);
+				if (idx > 0)
 				{
-					var questionMarkIndex = objectName.IndexOf(QUESTION_MARK);
-					objectName = questionMarkIndex >= 0 ? objectName.Substring(0, questionMarkIndex): objectName.Substring(0);
-					return provider;
+					providerObjectName = providerObjectName.Substring(0, idx);
 				}
+
+				// We store in DB, Path Encoded Urls. If the parameter is an absolute URL, we need to decode the ObjectName to get the real Object Name.
+				if (providerObjectName != null && PathUtil.IsAbsoluteUrl(objectNameOrUrl))
+					providerObjectName = StorageUtils.DecodeUrl(providerObjectName);
+				return true;
 			}
-			return null;
+			return false;
+		}
+
+		public static string GetProviderObjectAbsoluteUriSafe(ExternalProvider provider, string rawObjectUri)
+		{
+			string providerObjectName;
+			if (TryGetProviderObjectName(provider, rawObjectUri, out providerObjectName))
+			{
+				rawObjectUri = providerObjectName;
+			}
+			return rawObjectUri;
 		}
 	}
 }
