@@ -57,19 +57,19 @@ namespace TZ4Net
 		{
 			if (bitConverter==null)
 			{
-				throw new ArgumentNullException("bitConverter");
+				throw new ArgumentNullException(nameof(bitConverter));
 			}
 			if (stream==null)
 			{
-				throw new ArgumentNullException("stream");
+				throw new ArgumentNullException(nameof(stream));
 			}
 			if (encoding==null)
 			{
-				throw new ArgumentNullException("encoding");
+				throw new ArgumentNullException(nameof(encoding));
 			}
 			if (!stream.CanRead)
 			{
-				throw new ArgumentException("Stream isn't writable", "stream");
+				throw new ArgumentException("Stream isn't writable", nameof(stream));
 			}
 			this.stream = stream;
 			this.bitConverter = bitConverter;
@@ -296,15 +296,15 @@ namespace TZ4Net
 			CheckDisposed();
 			if (buffer==null)
 			{
-				throw new ArgumentNullException("buffer");
+				throw new ArgumentNullException(nameof(buffer));
 			}
 			if (index < 0)
 			{
-				throw new ArgumentOutOfRangeException("index");
+				throw new ArgumentOutOfRangeException(nameof(index));
 			}
 			if (count < 0)
 			{
-				throw new ArgumentOutOfRangeException("index");
+				throw new ArgumentOutOfRangeException(nameof(index));
 			}
 			if (count+index > buffer.Length)
 			{
@@ -370,15 +370,15 @@ namespace TZ4Net
 			CheckDisposed();
 			if (buffer==null)
 			{
-				throw new ArgumentNullException("buffer");
+				throw new ArgumentNullException(nameof(buffer));
 			}
 			if (index < 0)
 			{
-				throw new ArgumentOutOfRangeException("index");
+				throw new ArgumentOutOfRangeException(nameof(index));
 			}
 			if (count < 0)
 			{
-				throw new ArgumentOutOfRangeException("index");
+				throw new ArgumentOutOfRangeException(nameof(index));
 			}
 			if (count+index > buffer.Length)
 			{
@@ -412,7 +412,7 @@ namespace TZ4Net
 			CheckDisposed();
 			if (count < 0)
 			{
-				throw new ArgumentOutOfRangeException("count");
+				throw new ArgumentOutOfRangeException(nameof(count));
 			}
 			byte[] ret = new byte[count];
 			int index=0;
