@@ -226,11 +226,11 @@ namespace TZ4Net
 		{
 			if (value==null)
 			{
-				throw new ArgumentNullException("value");
+				throw new ArgumentNullException(nameof(value));
 			}
 			if (startIndex < 0 || startIndex > value.Length-bytesRequired)
 			{
-				throw new ArgumentOutOfRangeException("startIndex");
+				throw new ArgumentOutOfRangeException(nameof(startIndex));
 			}
 		}
 
@@ -493,7 +493,7 @@ namespace TZ4Net
 		{
 			if (buffer==null)
 			{
-				throw new ArgumentNullException("buffer", "Byte array must not be null");
+				throw new ArgumentNullException(nameof(buffer), "Byte array must not be null");
 			}
 			if (buffer.Length < index+bytes)
 			{
